requirements.txt
django>=4.2,<5.0
psycopg2-binary
python-decouple
dj-database-url
gunicorn
whitenoise
